requirements.txt
numpy==1.19.4
tensorflow-gpu==1.10.1
Keras==2.2.2
Soundfile==0.10.2
kapre==0.1.3.1
h5py==2.10.0
